requirements.txt
aniso8601==1.2.0
appdirs==1.4.0
click==6.7
Flask==0.12
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
SQLAlchemy==1.1.5
Werkzeug==0.11.15
dialogflow==0.5.1
google-api-core==1.4.1
